requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
rich>=13.7.0
PySocks>=1.7.1
stem>=1.8.2
aiohttp>=3.9.0
aiohttp-socks>=0.8.4
fake-useragent>=1.4.0
python-dateutil>=2.8.2
validators>=0.22.0
deep-translator>=1.11.0
PyYAML>=6.0.0
